refactor(main-api): JWT 토큰 관련 로직
