+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/jwt/tokens/AccessTokenSerializer.java
@@ -8,7 +8,6 @@ import com.whoz_in.domain.member.model.AccountType;
 import com.whoz_in.domain.member.model.MemberId;
 import com.whoz_in.main_api.shared.jwt.JwtProperties;
 import com.whoz_in.main_api.shared.jwt.JwtUtil;
-import com.whoz_in.main_api.shared.jwt.TokenType;
 import io.jsonwebtoken.Claims;
 import java.util.Map;
 import java.util.UUID;
