+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/jwt/tokens/OAuth2TempTokenSerializer.java
@@ -4,7 +4,6 @@ import static com.whoz_in.main_api.shared.jwt.JwtConst.OAUTH_USER_INFO_KEY;
 
 import com.whoz_in.main_api.shared.jwt.JwtProperties;
 import com.whoz_in.main_api.shared.jwt.JwtUtil;
-import com.whoz_in.main_api.shared.jwt.TokenType;
 import io.jsonwebtoken.Claims;
 import java.util.Map;
 import org.springframework.stereotype.Component;
