+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/jwt/tokens/TokenException.java
@@ -1,0 +1,9 @@
+package com.whoz_in.main_api.shared.jwt.tokens;
+
+import com.whoz_in.main_api.shared.application.ApplicationException;
+
+public class TokenException extends ApplicationException {
+    public TokenException(String errorCode, String message) {
+        super(errorCode, message);
+    }
+}
