+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/jwt/tokens/TokenSerializer.java
@@ -2,29 +2,37 @@ package com.whoz_in.main_api.shared.jwt.tokens;
 
 import com.whoz_in.main_api.shared.jwt.JwtProperties;
 import com.whoz_in.main_api.shared.jwt.JwtUtil;
-import com.whoz_in.main_api.shared.jwt.TokenType;
 import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.JwtException;
 import java.util.Map;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
 /*
 문자열 토큰 <-> 토큰 VO 변환을 지원한다.
 문자열 토큰 -> 토큰 VO 변환 시 올바른 토큰인지 검증할 책임이 있다.
  */
+@Slf4j
 @Component
 @RequiredArgsConstructor
 public abstract class TokenSerializer<T extends Token> {
     private final JwtUtil jwtUtil;
     private final JwtProperties jwtProperties;
 
-    public final T deserialize(String token){
-        Claims claims = jwtUtil.getClaims(token);
-        validate(claims);
-        return buildToken(claims);
+    public final Optional<T> deserialize(String token){
+        try {
+            Claims claims = jwtUtil.getClaims(token);
+            validate(claims);
+            return Optional.of(buildToken(claims));
+        } catch (JwtException e) {
+            log.warn("잘못된 토큰: {}", e.getMessage());
+            return Optional.empty();
+        }
     }
 
-    public final String serialize(T jwtInfo){
+    public final String serialize(T jwtInfo) {
         Map<String, String> claims = buildClaims(jwtInfo);
         return jwtUtil.createJwt(getTokenType(), claims, jwtProperties.getTokenExpiry(getTokenType()));
     }
